Ligther Ancestor

+++ examples/mnist_conv2d.mojo
@@ -166,7 +166,7 @@ fn train_mnist() raises:
             loss.backward()
             optimizer.step()
 
-            train_loss += loss.item() * batch.batch_size
+            train_loss += loss.item() * Float32(batch.batch_size)
             train_correct += compute_accuracy(pred, batch.labels)
             train_total += batch.batch_size
             batch_num += 1
@@ -185,16 +185,16 @@ fn train_mnist() raises:
             var pred = model(batch.features)
             var loss = criterion(pred, batch.labels)
 
-            val_loss += loss.item() * batch.batch_size
+            val_loss += loss.item() * Float32(batch.batch_size)
             val_correct += compute_accuracy(pred, batch.labels)
             val_total += batch.batch_size
 
         # --- Epoch Report ---
         var epoch_time = now() - epoch_start
-        var avg_train_loss = train_loss / train_total
-        var train_acc = 100.0 * train_correct / train_total
-        var avg_val_loss = val_loss / val_total
-        var val_acc = 100.0 * val_correct / val_total
+        var avg_train_loss = train_loss / Float32(train_total)
+        var train_acc = 100.0 * Float64(train_correct) / Float64(train_total)
+        var avg_val_loss = val_loss / Float32(val_total)
+        var val_acc = 100.0 * Float64(val_correct) / Float64(val_total)
         print(
             "Epoch",
             epoch + 1,
